BW-28 Fixing Loading Issues
